+++ test/textlint-filter-rule-comments-fix-test.js
@@ -2,16 +2,16 @@
 "use strict";
 const TextLintCore = require("textlint").TextLintCore;
 const TextLintNodeType = require("textlint").TextLintNodeType;
-const ignoreRule = require("../src/textlint-rule-ignore-comments");
+const filterRule = require("../src/textlint-filter-rule-comments");
 const reportRule = require("textlint-rule-report-node-types");
 const assert = require("power-assert");
-describe("textlint-rule-ignore-node-types --fix", function () {
+describe.skip("textlint-rule-ignore-node-types --fix", function () {
     context("no options", function () {
         context("when before textlint-enable", function () {
             it("should not ignored", function () {
                 const textlint = new TextLintCore();
                 textlint.setupRules({
-                    ignore: ignoreRule,
+                    ignore: filterRule,
                     report: reportRule
                 }, {
                     report: {
@@ -36,7 +36,7 @@ This is ignored.
             it("should messages is ignored between disable and enable", function () {
                 const textlint = new TextLintCore();
                 textlint.setupRules({
-                    ignore: ignoreRule,
+                    ignore: filterRule,
                     report: reportRule
                 }, {
                     report: {
@@ -58,7 +58,7 @@ This is text.
             it("should not ignored", function () {
                 const textlint = new TextLintCore();
                 textlint.setupRules({
-                    ignore: ignoreRule,
+                    ignore: filterRule,
                     report: reportRule
                 }, {
                     report: {
@@ -85,7 +85,7 @@ This is Error.
             it("should ignore messages of ruleA", function () {
                 const textlint = new TextLintCore();
                 textlint.setupRules({
-                    ignore: ignoreRule,
+                    ignore: filterRule,
                     ruleA: reportRule
                 }, {
                     ruleA: {
@@ -105,7 +105,7 @@ This is text.
             it("should not ignore messages of other rules", function () {
                 const textlint = new TextLintCore();
                 textlint.setupRules({
-                    ignore: ignoreRule,
+                    ignore: filterRule,
                     ruleX: reportRule
                 }, {
                     ruleX: {
@@ -127,7 +127,7 @@ This is text.
             it("should ignore messages of ruleA", function () {
                 const textlint = new TextLintCore();
                 textlint.setupRules({
-                    ignore: ignoreRule,
+                    ignore: filterRule,
                     ruleA: reportRule,
                     ruleB: reportRule
                 }, {
