+++ test/textlint-filter-rule-comments-test.js
@@ -2,7 +2,7 @@
 "use strict";
 const TextLintCore = require("textlint").TextLintCore;
 const TextLintNodeType = require("textlint").TextLintNodeType;
-const ignoreRule = require("../src/textlint-rule-ignore-comments");
+const filterRule = require("../src/textlint-filter-rule-comments");
 const reportRule = require("textlint-rule-report-node-types");
 const assert = require("power-assert");
 describe("textlint-rule-ignore-node-types", function () {
@@ -11,7 +11,7 @@ describe("textlint-rule-ignore-node-types", function () {
             it("should not ignored", function () {
                 const textlint = new TextLintCore();
                 textlint.setupRules({
-                    ignore: ignoreRule,
+                    ignore: filterRule,
                     report: reportRule
                 }, {
                     report: {
@@ -35,7 +35,7 @@ This is ignored.
             it("should messages is ignored between disable and enable", function () {
                 const textlint = new TextLintCore();
                 textlint.setupRules({
-                    ignore: ignoreRule,
+                    ignore: filterRule,
                     report: reportRule
                 }, {
                     report: {
@@ -57,7 +57,7 @@ This is text.
             it("should not ignored", function () {
                 const textlint = new TextLintCore();
                 textlint.setupRules({
-                    ignore: ignoreRule,
+                    ignore: filterRule,
                     report: reportRule
                 }, {
                     report: {
@@ -84,7 +84,7 @@ This is Error.
             it("should ignore messages of ruleA", function () {
                 const textlint = new TextLintCore();
                 textlint.setupRules({
-                    ignore: ignoreRule,
+                    ignore: filterRule,
                     ruleA: reportRule
                 }, {
                     ruleA: {
@@ -104,7 +104,7 @@ This is text.
             it("should not ignore messages of other rules", function () {
                 const textlint = new TextLintCore();
                 textlint.setupRules({
-                    ignore: ignoreRule,
+                    ignore: filterRule,
                     ruleX: reportRule
                 }, {
                     ruleX: {
@@ -126,7 +126,7 @@ This is text.
             it("should ignore messages of ruleA", function () {
                 const textlint = new TextLintCore();
                 textlint.setupRules({
-                    ignore: ignoreRule,
+                    ignore: filterRule,
                     ruleA: reportRule,
                     ruleB: reportRule
                 }, {
